# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Skills4Attendance API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tutors:
    get:
      operationId: listTutors
      tags: [tutors]
      parameters:
        - name: active
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: List of tutors
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Tutor" }
    post:
      operationId: createTutor
      tags: [tutors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutorInput"
      responses:
        "201":
          description: Tutor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tutor"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tutors/{id}:
    get:
      operationId: getTutor
      tags: [tutors]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Tutor detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tutor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTutor
      tags: [tutors]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutorUpdate"
      responses:
        "200":
          description: Tutor updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tutor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------- Learners ----------
...
  /reports/learner/{id}:
    get:
      operationId: getLearnerReport
      tags: [reports]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Learner attendance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearnerReport"
  /reports/cohort/{id}:
    get:
      operationId: getCohortReport
      tags: [reports]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Cohort attendance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CohortReport"
  /reports/tutor/{id}:
    get:
      operationId: getTutorReport
      tags: [reports]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Tutor attendance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorReport"
  /reports/organisation:
    get:
      operationId: getOrganisationReport
      tags: [reports]
      parameters:
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
        - name: programme
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Organisation-wide attendance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganisationReport"
  /reports/programme:
    get:
      operationId: getProgrammeReport
      tags: [reports]
      parameters:
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Attendance breakdown by programme
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProgrammeAttendanceRow" }
  /reports/export:
    get:
      operationId: exportReport
      tags: [reports]
      parameters:
        - name: reportType
          in: query
          required: true
          schema:
            type: string
            enum: [learner, cohort, tutor, organisation, programme, allocation-history]
        - name: entityId
          in: query
          schema: { type: integer }
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
        - name: tutorId
          in: query
          schema: { type: integer }
        - name: cohortId
          in: query
          schema: { type: integer }
        - name: programme
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/LearnerStatus" }
      responses:
        "200":
          description: CSV export content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsvContent"

  # ---------- Audit ----------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    LearnerStatus:
      type: string
      enum: [active, withdrawn, completed, paused]
...
    DeliveryDay:
      type: string
      enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
...
    Tutor:
      type: object
      required:
        [
          id,
          userId,
          firstName,
          lastName,
          email,
          employeeRef,
          active,
          externalSystemId,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: integer }
        userId: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        employeeRef: { type: string }
        active: { type: boolean }
        externalSystemId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TutorInput:
      type: object
      required: [firstName, lastName, email, password, employeeRef]
      properties:
        firstName: { type: string, minLength: 1 }
        lastName: { type: string, minLength: 1 }
        email: { type: string, minLength: 1 }
        password: { type: string, minLength: 8 }
        employeeRef: { type: string, minLength: 1 }
        active: { type: boolean, default: true }
        externalSystemId: { type: string }
    TutorUpdate:
      type: object
      properties:
        firstName: { type: string, minLength: 1 }
        lastName: { type: string, minLength: 1 }
        email: { type: string, minLength: 1 }
        password: { type: string, minLength: 8 }
        employeeRef: { type: string, minLength: 1 }
        active: { type: boolean }
        externalSystemId: { type: string }
    Learner:
      type: object
      required:
        [
          id,
          learnerRef,
          uln,
          firstName,
          lastName,
          email,
          employer,
          programme,
          level,
          startDate,
          plannedEndDate,
          status,
          tutorId,
          cohortId,
          tutorName,
          cohortName,
          externalSystemId,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: integer }
        learnerRef: { type: string }
        uln: { type: ["string", "null"] }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: ["string", "null"] }
        employer: { type: ["string", "null"] }
        programme: { type: string }
        level: { type: string }
        startDate: { type: string, format: date }
        plannedEndDate: { type: ["string", "null"], format: date }
        status: { $ref: "#/components/schemas/LearnerStatus" }
        tutorId: { type: ["integer", "null"] }
        cohortId: { type: ["integer", "null"] }
        tutorName: { type: ["string", "null"] }
        cohortName: { type: ["string", "null"] }
        externalSystemId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CsvContent:
      type: object
      required: [csv]
      properties:
        csv: { type: string }
        filename: { type: string }
...
    Cohort:
      type: object
      required:
        [
          id,
          name,
          programme,
          level,
          tutorId,
          tutorName,
          deliveryDay,
          sessionStartTime,
          sessionEndTime,
          startDate,
          endDate,
          active,
          externalSystemId,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: integer }
        name: { type: string }
        programme: { type: string }
        level: { type: string }
        tutorId: { type: ["integer", "null"] }
        tutorName: { type: ["string", "null"] }
        deliveryDay: { $ref: "#/components/schemas/DeliveryDay" }
        sessionStartTime: { type: string }
        sessionEndTime: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: ["string", "null"], format: date }
        active: { type: boolean }
        externalSystemId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    AttendanceTotals:
      type: object
      required:
        [
          scheduledHours,
          attendedHours,
          authorisedAbsenceHours,
          unauthorisedAbsenceHours,
          lateCount,
          sessionCount,
          attendancePercentage,
        ]
      properties:
        scheduledHours: { type: number }
        attendedHours: { type: number }
        authorisedAbsenceHours: { type: number }
        unauthorisedAbsenceHours: { type: number }
        lateCount: { type: integer }
        sessionCount: { type: integer }
        attendancePercentage: { type: number }
    LearnerAttendanceRow:
      type: object
      required: [learnerId, learnerName, learnerRef, totals]
      properties:
        learnerId: { type: integer }
        learnerName: { type: string }
        learnerRef: { type: string }
        totals: { $ref: "#/components/schemas/AttendanceTotals" }
    CohortAttendanceRow:
      type: object
      required: [cohortId, cohortName, totals]
      properties:
        cohortId: { type: integer }
        cohortName: { type: string }
        totals: { $ref: "#/components/schemas/AttendanceTotals" }
    ProgrammeAttendanceRow:
      type: object
      required: [programme, totals]
      properties:
        programme: { type: string }
        totals: { $ref: "#/components/schemas/AttendanceTotals" }
    LearnerReport:
      type: object
      required: [learner, totals]
      properties:
        learner: { $ref: "#/components/schemas/Learner" }
        totals: { $ref: "#/components/schemas/AttendanceTotals" }
    CohortReport:
      type: object
      required: [cohort, totals, learnerBreakdown]
      properties:
        cohort: { $ref: "#/components/schemas/Cohort" }
        totals: { $ref: "#/components/schemas/AttendanceTotals" }
        learnerBreakdown:
          type: array
          items: { $ref: "#/components/schemas/LearnerAttendanceRow" }
    TutorReport:
      type: object
      required: [tutor, totals, cohortBreakdown]
      properties:
        tutor: { $ref: "#/components/schemas/Tutor" }
        totals: { $ref: "#/components/schemas/AttendanceTotals" }
        cohortBreakdown:
          type: array
          items: { $ref: "#/components/schemas/CohortAttendanceRow" }
    OrganisationReport:
      type: object
      required: [totals, programmeBreakdown, cohortBreakdown]
      properties:
        totals: { $ref: "#/components/schemas/AttendanceTotals" }
        programmeBreakdown:
          type: array
          items: { $ref: "#/components/schemas/ProgrammeAttendanceRow" }
        cohortBreakdown:
          type: array
          items: { $ref: "#/components/schemas/CohortAttendanceRow" }
...